# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Enhanced RAG Chatbot API",
    "version": "2.0.0",
    "description": "Comprehensive API documentation for the Enhanced RAG Chatbot with JWT Authentication, Dual Database Architecture, and Free AI Components",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    },
    {
      "url": "https://api.yourapp.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication. Format: 'Bearer <token>'"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user identifier"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "isVerified": {
            "type": "boolean",
            "description": "Email verification status"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Account creation timestamp"
          }
        }
      },
...
      "SystemStats": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["healthy", "degraded", "unhealthy"]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "string",
            "description": "API version"
          },
          "components": {
            "type": "object",
            "properties": {
              "database": {
                "type": "object",
                "properties": {
                  "supabase": {
                    "type": "boolean"
                  },
                  "inMemory": {
                    "type": "boolean"
                  }
                }
              },
              "vectorStore": {
                "type": "object",
                "properties": {
                  "chromadb": {
                    "type": "boolean"
                  },
                  "collections": {
                    "type": "integer"
                  }
                }
              },
              "embeddings": {
                "type": "object",
                "properties": {
                  "service": {
                    "type": "string"
                  },
                  "model": {
                    "type": "string"
                  },
                  "initialized": {
                    "type": "boolean"
                  }
                }
              },
              "llm": {
                "type": "object",
                "properties": {
                  "models": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "currentModel": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "statistics": {
            "type": "object",
            "properties": {
              "totalDocuments": {
                "type": "integer"
              },
              "totalChunks": {
                "type": "integer"
              },
              "totalQueries": {
                "type": "integer"
              },
              "averageResponseTime": {
                "type": "number"
              },
              "cacheHitRate": {
                "type": "number"
              }
            }
          },
          "features": {
            "type": "object",
            "properties": {
              "freeAndOpenSource": {
                "type": "boolean"
              },
              "localEmbeddings": {
                "type": "boolean"
              },
              "contextMemory": {
                "type": "boolean"
              },
              "multiModelLLM": {
                "type": "boolean"
              },
              "vectorSearch": {
                "type": "boolean"
              },
              "semanticCache": {
                "type": "boolean"
              },
              "smartChunking": {
                "type": "boolean"
              },
              "documentSummaries": {
                "type": "boolean"
              }
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message"
          },
          "code": {
            "type": "integer",
            "description": "Error code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "paths": {
...
    "/user/profile": {
      "get": {
        "summary": "Get user profile information",
        "tags": ["User"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/stats": {
      "get": {
        "summary": "Get comprehensive system statistics and health",
        "tags": ["System"],
        "responses": {
          "200": {
            "description": "System statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStats"
                }
              }
            }
          },
          "500": {
            "description": "Failed to retrieve system stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}